moved settings and gallery completely into helper files. (flet ui)
