+++ webui/flet/scripts/flet_file_manager.py
@@ -1,0 +1,8 @@
+# flet_file_manager.py
+
+# Flet imports
+import flet as ft
+
+# utils imports
+from scripts import flet_utils
+
